# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API with Swagger",
        "description": "API Documentation User",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://103.146.202.172:8090/preview/api-laravel-swagger.my.id/"
        },
        {
            "url": "/preview/api-laravel-swagger.my.id"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "e10957ea135d6b0551c4851a13a56c88",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "operationId": "efc3fab9e157b839e53438f9c9819c9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "dani"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "dani@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "123 Street"
                                    },
                                    "no_tlp": {
                                        "type": "string",
                                        "example": "08123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by ID",
                "operationId": "a58b2476400765004bf1868f8d787e0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/users/delete/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by ID",
                "operationId": "75ca25ebd02af4ba58706fccb7ae8179",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/users/update/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user by ID",
                "operationId": "97d6cee3e6d901306b2a691941b07ef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "dani"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "dani@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "123 Street"
                                    },
                                    "no_tlp": {
                                        "type": "string",
                                        "example": "08123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "dani"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test1@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Test No. 123"
                    },
                    "no_tlp": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ],
    "externalDocs": {
        "description": "link to repository danibahri/api-laravel-swagger",
        "url": "https://github.com/danibahri/api-laravel-swagger"
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}